feat: add support for TMA method

+++ mperf-data/src/ipc.rs
@@ -4,7 +4,7 @@ use crate::Event;
 
 #[derive(Encode, Decode, Clone, Debug)]
 pub struct IPCString {
-    pub key: u64,
+    pub key: u128,
     pub value: String,
 }
